update crud & login

+++ app/Http/Controllers/SkemaController.php
@@ -25,13 +25,13 @@ class SkemaController extends Controller
             'jenis_standar' => ['required','in:KKNI,Okupasi,Klaster'],
             'kode_skema' => ['required','unique:m_skema,kode_skema'],
             'judul_skema' => ['required', 'string', 'max:255'],
-            'event_id' => ['required']
+            'event_id' => ['required','exists:m_event,uuid'],
         ], $this->messageValidation());
 
         if ($validator->fails()) {
             return response()->json(['status' => 'error', 'message' => $validator->errors()], 422);
         }
-        $validated = $validator->validated();
+        $validated = $validator->validate();
         $event = Event::where('uuid', $validated['event_id'])->first();
         if(empty($event)) {
             return response()->json(['status' => 'error', 'message' => 'Data event tidak ditemukan'], 404);
@@ -68,13 +68,13 @@ class SkemaController extends Controller
             'jenis_standar' => ['required','in:KKNI,Okupasi,Klaster'],
             'kode_skema' => ['required','unique:m_skema,kode_skema,'. $uuid . ',uuid'],
             'judul_skema' => ['required', 'string', 'max:255'],
-            'event_id' => ['required']
+            'event_id' => ['required','exists:m_event,uuid'],
         ], $this->messageValidation());
 
         if ($validator->fails()) {
             return response()->json(['status' => 'error', 'message' => $validator->errors()], 422);
         }
-        $validated = $validator->validated();
+        $validated = $validator->validate();
         $event = Event::where('uuid', $validated['event_id'])->first();
         if(empty($event)) {
             return response()->json(['status' => 'error', 'message' => 'Data event tidak ditemukan'], 404);
@@ -99,10 +99,13 @@ class SkemaController extends Controller
     public function destroy($uuid)
     {
         $skema = Skema::where('uuid', $uuid);
+        if(empty($skema->first())) {
+            return response()->json(['status' => 'error', 'message' => 'Data skema tidak ditemukan'], 404);
+        }
         if($skema->delete()) {
             return response()->json(['status' => 'success','message' => 'Data skema berhasil dihapus'], 200);
         } else {
-            return response()->json(['status' => 'error', 'message' => 'Data skema tidak ditemukan'], 404);
+            return response()->json(['status' => 'error', 'message' => 'Data skema gagal dihapus'], 500);
         }
     }
